requirements.txt
# Security packages
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
email-validator==2.1.0

alembic==1.14.1
annotated-types==0.7.0
argcomplete==1.10.3
Authlib==1.4.1
beautifulsoup4==4.8.2
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
compressed_rtf==1.0.6
cryptography==44.0.0
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-vision==3.10.1
googleapis-common-protos==1.66.0
huggingface_hub==0.27.1
PyPDF2==3.0.1
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
IMAPClient==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.1
Mako==1.3.9
MarkupSafe==3.0.2
oauthlib==3.2.2
olefile==0.47
packaging==24.2
pdfminer.six==20191110
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==0.6.23
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.8.1
SQLAlchemy==2.0.38
textract==1.6.5
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
xlrd==1.2.0
XlsxWriter==3.2.2
